Fixed a warning.

+++ server/src/TPlayerProps.cpp
@@ -710,7 +710,7 @@ void TPlayer::setProps(CString& pPacket, bool pForward, bool pForwardToSelf, TPl
 				if ((short)len & 0x0001) z2 = -z2;
 
 				// Let pre-2.30+ clients see 2.30+ movement.
-				z2 -= (float)(25 * 16);
+				z2 -= 25 * 16;
 				z = (float)z2 / 16.0f;
 				levelBuff2 >> (char)PLPROP_Z << getProp(PLPROP_Z);
 				break;
